Move account header and avatar methods to a concern
